requirements-dev.txt
-e .
ruff~=0.11.0
pytest~=8.3.0
pytest-cov~=6.0.0
python-semantic-release~=8.0.0
build~=1.0
twine~=6.1.0
pre-commit>=4.0.0
yamllint~=1.35.0
bandit~=1.9.0
